move syntax/file-ext logic to another file
